cleanup/shuffle
- toggle action states
- eventfilter
- undostack and undocommands

falling back to some QGraphicsview actions, such as move.
this "works" but not clear how/where to create undoCommands.

+++ qucs/qucs/netlist.cpp
@@ -25,7 +25,7 @@ Net* NetList::newNet()
 	}else{ untested();
 		idx = _g.top();
 		_g.pop();
-		assert(_l[idx]->size() == -1);
+		assert(_l[idx]->size() == size_t(-1));
 		_l[idx]->inc_nodes();
 		assert(_l[idx]->getPos() == idx);
 	}
